menu

+++ coffeeshop/src/main/java/com/coffeeshop/DAO/ShopDAO.java
@@ -6,5 +6,5 @@ public interface ShopDAO {
     public ShopEntity create(ShopEntity shopEntity);
     public ShopEntity Update(ShopEntity shopEntity);
     public ShopEntity Delete(ShopEntity shopEntity);
-    public ShopEntity getShop(int shopId);
+    public ShopEntity getShop(long shopId);
 }
